# openapi.yaml
...
openapi: 3.0.0
info:
  title: GigaWrks User Service
  description: This document contains the contract of the User service with all the REST APIs needed for the users with support to fetch available countries.
  contact:
    email: nehul.rangappa@gmail.com
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/NEHULRANGAPPA_1/gigawrks-user-service/1.0.0
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /countries:
    get:
      tags:
      - Countries
      summary: Fetch all the countries
      description: Fetch all the available countries from the external client
      operationId: getCountries
      parameters:
      - name: id
        in: query
        description: Identifier for finding the appropriate country
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: code
        in: query
        description: Country Code for finding the appropriate country
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: name
        in: query
        description: Country Name for finding the appropriate country
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Countries fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/countriesOutput'
        "400":
          description: Bad Request
        "404":
          description: No countries found
        "500":
          description: "Internal Server Error: Please try again"
components:
  schemas:
...
    countriesOutput:
      type: array
      items:
        type: object
        properties:
          id:
            type: integer
            example: 1
          commonName:
            type: string
            example: United States
          officialName:
            type: string
            example: United States of America
          countryCode:
            type: string
            example: US
          capital:
            type: string
          region:
            type: string
          subregion:
            type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT